requirements.txt
langchain
langchain-community
langchain-openai
fastapi
python-dotenv
pydantic
faiss-cpu
tiktoken
PyYAML
PyMuPDF
pdf2image
pytesseract
Pillow
docling
pdfplumber
scikit-learn
nltk
numpy
ragas
datasets
opencensus-ext-azure
opencensus-ext-requests
semantic-kernel
azure-search-documents
azure-identity
azure-ai-ml
azure-cognitiveservices-language-textanalytics
azure-core
openai 